# Cython-based implementation of reader module (~25% faster)