--- a/fastavro/_reader.pyx
+++ b/fastavro/_reader.pyx
@@ -13,7 +13,7 @@ from decimal import localcontext, Decimal
 from uuid import UUID
 
 from ._six import (
-    MemoryIO, xrange, btou, utob, iteritems, is_str, str2ints, fstint
+    xrange, btou, utob, iteritems, is_str, str2ints, fstint
 )
 from ._schema import (
     extract_record_type, acquaint_schema, populate_schema_defs,
@@ -119,12 +119,12 @@ cpdef match_schemas(w_schema, r_schema):
         raise SchemaResolutionError(error_msg)
 
 
-cpdef inline read_null(object fo, writer_schema=None, reader_schema=None):
+cpdef inline read_null(ReaderBase fo, writer_schema=None, reader_schema=None):
     """null is written as zero bytes."""
     return None
 
 
-cpdef inline bint read_boolean(object fo, writer_schema=None, reader_schema=None) except? 2:
+cpdef inline bint read_boolean(ReaderBase fo, writer_schema=None, reader_schema=None) except? 2:
     """A boolean is written as a single byte whose value is either 0 (false) or
     1 (true).
     """
@@ -215,7 +215,7 @@ cpdef _read_decimal(data, size, writer_schema):
     return scaled_datum
 
 
-cpdef long read_long(fo, writer_schema=None, reader_schema=None) except? -1:
+cpdef long read_long(ReaderBase fo, writer_schema=None, reader_schema=None) except? -1:
     """int and long values are written using variable-length, zig-zag
     coding."""
     cdef unsigned long b
@@ -245,7 +245,7 @@ cdef union float_int:
     unsigned int n
 
 
-cpdef float read_float(fo, writer_schema=None, reader_schema=None) except? -1.0:
+cpdef float read_float(ReaderBase fo, writer_schema=None, reader_schema=None) except? -1.0:
     """A float is written as 4 bytes.
 
     The float is converted into a 32-bit integer using a method equivalent to
@@ -269,7 +269,7 @@ cdef union double_long:
     unsigned long n
 
 
-cpdef double read_double(fo, writer_schema=None, reader_schema=None) except? -1.0:
+cpdef double read_double(ReaderBase fo, writer_schema=None, reader_schema=None) except? -1.0:
     """A double is written as 8 bytes.
 
     The double is converted into a 64-bit integer using a method equivalent to
@@ -294,26 +294,26 @@ cpdef double read_double(fo, writer_schema=None, reader_schema=None) except? -1.
         raise ReadError
 
 
-cpdef read_bytes(fo, writer_schema=None, reader_schema=None):
+cpdef read_bytes(ReaderBase fo, writer_schema=None, reader_schema=None):
     """Bytes are encoded as a long followed by that many bytes of data."""
     cdef long size = read_long(fo)
     return fo.read(size)
 
 
-cpdef unicode read_utf8(fo, writer_schema=None, reader_schema=None):
+cpdef unicode read_utf8(ReaderBase fo, writer_schema=None, reader_schema=None):
     """A string is encoded as a long followed by that many bytes of UTF-8
     encoded character data.
     """
     return btou(read_bytes(fo), 'utf-8')
 
 
-cpdef read_fixed(fo, writer_schema, reader_schema=None):
+cpdef read_fixed(ReaderBase fo, writer_schema, reader_schema=None):
     """Fixed instances are encoded using the number of bytes declared in the
     schema."""
     return fo.read(writer_schema['size'])
 
 
-cpdef read_enum(fo, writer_schema, reader_schema=None):
+cpdef read_enum(ReaderBase fo, writer_schema, reader_schema=None):
     """An enum is encoded by a int, representing the zero-based position of the
     symbol in the schema.
     """
@@ -326,7 +326,7 @@ cpdef read_enum(fo, writer_schema, reader_schema=None):
     return symbol
 
 
-cpdef read_array(fo, writer_schema, reader_schema=None):
+cpdef read_array(ReaderBase fo, writer_schema, reader_schema=None):
     """Arrays are encoded as a series of blocks.
 
     Each block consists of a long count value, followed by that many array
@@ -352,16 +352,16 @@ cpdef read_array(fo, writer_schema, reader_schema=None):
 
         if reader_schema:
             for i in xrange(block_count):
-                read_items.append(read_data(fo, writer_schema['items'], reader_schema['items']))
+                read_items.append(_read_data(fo, writer_schema['items'], reader_schema['items']))
         else:
             for i in xrange(block_count):
-                read_items.append(read_data(fo, writer_schema['items']))
+                read_items.append(_read_data(fo, writer_schema['items']))
         block_count = read_long(fo)
 
     return read_items
 
 
-cpdef read_map(fo, writer_schema, reader_schema=None):
+cpdef read_map(ReaderBase fo, writer_schema, reader_schema=None):
     """Maps are encoded as a series of blocks.
 
     Each block consists of a long count value, followed by that many key/value
@@ -387,17 +387,17 @@ cpdef read_map(fo, writer_schema, reader_schema=None):
         if reader_schema:
             for i in xrange(block_count):
                 key = read_utf8(fo)
-                read_items[key] = read_data(fo, writer_schema['values'], reader_schema['values'])
+                read_items[key] = _read_data(fo, writer_schema['values'], reader_schema['values'])
         else:
             for i in xrange(block_count):
                 key = read_utf8(fo)
-                read_items[key] = read_data(fo, writer_schema['values'])
+                read_items[key] = _read_data(fo, writer_schema['values'])
         block_count = read_long(fo)
 
     return read_items
 
 
-cpdef read_union(fo, writer_schema, reader_schema=None):
+cpdef read_union(ReaderBase fo, writer_schema, reader_schema=None):
     """A union is encoded by first writing a long value indicating the
     zero-based position within the union of the schema of its value.
 
@@ -409,19 +409,19 @@ cpdef read_union(fo, writer_schema, reader_schema=None):
         # Handle case where the reader schema is just a single type (not union)
         if not isinstance(reader_schema, list):
             if match_types(writer_schema[index], reader_schema):
-                return read_data(fo, writer_schema[index], reader_schema)
+                return _read_data(fo, writer_schema[index], reader_schema)
         else:
             for schema in reader_schema:
                 if match_types(writer_schema[index], schema):
-                    return read_data(fo, writer_schema[index], schema)
+                    return _read_data(fo, writer_schema[index], schema)
         msg = 'schema mismatch: %s not found in %s' % \
             (writer_schema, reader_schema)
         raise SchemaResolutionError(msg)
     else:
-        return read_data(fo, writer_schema[index])
+        return _read_data(fo, writer_schema[index])
 
 
-cpdef read_record(fo, writer_schema, reader_schema=None):
+cpdef read_record(ReaderBase fo, writer_schema, reader_schema=None):
     """A record is encoded by encoding the values of its fields in the order
     that they are declared. In other words, a record is encoded as just the
     concatenation of the encodings of its fields.  Field values are encoded per
@@ -443,18 +443,18 @@ cpdef read_record(fo, writer_schema, reader_schema=None):
     record = {}
     if reader_schema is None:
         for field in writer_schema['fields']:
-            record[field['name']] = read_data(fo, field['type'])
+            record[field['name']] = _read_data(fo, field['type'])
     else:
         readers_field_dict = {f['name']: f for f in reader_schema['fields']}
         for field in writer_schema['fields']:
             readers_field = readers_field_dict.get(field['name'])
             if readers_field:
-                record[field['name']] = read_data(fo,
+                record[field['name']] = _read_data(fo,
                                                   field['type'],
                                                   readers_field['type'])
             else:
                 # should implement skip
-                read_data(fo, field['type'], field['type'])
+                _read_data(fo, field['type'], field['type'])
 
         # fill in default values
         if len(readers_field_dict) > len(record):
@@ -504,6 +504,10 @@ READERS = {
 
 
 cpdef read_data(fo, writer_schema, reader_schema=None):
+    return _read_data(FileObjectReader(fo), writer_schema, reader_schema)
+
+
+cpdef _read_data(ReaderBase fo, writer_schema, reader_schema=None):
     """Read data from file object according to schema."""
 
     record_type = extract_record_type(writer_schema)
@@ -522,24 +526,55 @@ cpdef read_data(fo, writer_schema, reader_schema=None):
         raise EOFError('cannot read %s from %s' % (record_type, fo))
 
 
-cpdef skip_sync(fo, sync_marker):
+cpdef skip_sync(ReaderBase fo, sync_marker):
     """Skip an expected sync marker, complaining if it doesn't match"""
     if fo.read(SYNC_SIZE) != sync_marker:
         raise ValueError('expected sync marker not found')
 
 
-cpdef null_read_block(fo):
+cdef class ReaderBase(object):
+    cpdef bytes read(self, int size):
+        raise NotImplemented
+
+
+cdef class MemoryReader(ReaderBase):
+    cdef bytes data
+    cdef unsigned long position
+
+    def __init__(self, data):
+        self.data = data
+        self.position = 0
+
+    cpdef bytes read(self, int size):
+        cdef bytes result
+        result = self.data[self.position : self.position + size]
+        self.position += size
+        return result
+
+
+cdef class FileObjectReader(ReaderBase):
+    cdef object fo
+
+    def __init__(self, fo):
+        self.fo = fo
+
+    cpdef bytes read(self, int size):
+        return self.fo.read(size)
+
+
+cpdef MemoryReader null_read_block(fo):
     """Read block in "null" codec."""
-    read_long(fo, None)
-    return fo
+    length = read_long(fo, None)
+    data = fo.read(length)
+    return MemoryReader(data)
 
 
-cpdef deflate_read_block(fo):
+cpdef MemoryReader deflate_read_block(fo):
     """Read block in "deflate" codec."""
     data = read_bytes(fo, None)
     # -15 is the log of the window size; negative indicates "raw" (no
     # zlib headers) decompression.  See zlib.h.
-    return MemoryIO(decompress(data, -15))
+    return MemoryReader(decompress(data, -15))
 
 
 BLOCK_READERS = {
@@ -555,15 +590,17 @@ except ImportError:
     pass
 
 
-cpdef snappy_read_block(fo):
+cpdef MemoryReader snappy_read_block(fo):
     length = read_long(fo, None)
     data = fo.read(length - 4)
     fo.read(4)  # CRC
-    return MemoryIO(snappy.decompress(data))
+    return MemoryReader(snappy.decompress(data))
 
 
-def _iter_avro(fo, header, codec, writer_schema, reader_schema):
+def _iter_avro(ReaderBase fo, header, codec, writer_schema, reader_schema):
     """Return iterator over avro records."""
+    cdef ReaderBase block_fo
+
     sync_marker = header['sync']
     # Value in schema is bytes
 
@@ -577,7 +614,7 @@ def _iter_avro(fo, header, codec, writer_schema, reader_schema):
         block_fo = read_block(fo)
 
         for i in xrange(block_count):
-            yield read_data(block_fo, writer_schema, reader_schema)
+            yield _read_data(block_fo, writer_schema, reader_schema)
 
         skip_sync(fo, sync_marker)
 
@@ -604,7 +641,8 @@ class iter_avro:
         """
         self.fo = fo
         try:
-            self._header = read_data(fo, HEADER_SCHEMA)
+            fo_reader = FileObjectReader(fo)
+            self._header = _read_data(fo_reader, HEADER_SCHEMA)
         except StopIteration:
             raise ValueError('cannot read header - is it an avro file?')
 
@@ -621,7 +659,7 @@ class iter_avro:
         acquaint_schema(self.writer_schema, READERS)
         if reader_schema:
             populate_schema_defs(reader_schema)
-        self._records = _iter_avro(fo,
+        self._records = _iter_avro(fo_reader,
                                    self._header,
                                    self.codec,
                                    self.writer_schema,