… fastavro/_reader_common.py
… fastavro/_writer.pyx
… setup.py
… tests/test_fastavro.py