This PR speeds up `_reader.pyx` by applying similar Cython techniques as those seen in `_writer.pyx`. There is a definite speed improvement (about 25%), although less dramatic than for the writer.

Here are some benchmark numbers using the 10,000-record cases from the `benchmark.py` script @scottbelden had provided.

```
Previous release:
Writing 10000 records to one file...
... 10 runs averaged 0.0102419137955 seconds
Reading 10000 records from one file...
... 10 runs averaged 0.0207515954971 seconds

Writing 10000 records to one file...
... 10 runs averaged 0.0623118400574 seconds
Reading 10000 records from one file...
... 10 runs averaged 0.121334815025 seconds

==

This PR:
Writing 10000 records to one file...
... 50 runs averaged 0.00834895133972 seconds
Reading 10000 records from one file...
... 50 runs averaged 0.0154855728149 seconds

Writing 10000 records to one file...
... 50 runs averaged 0.0528964567184 seconds
Reading 10000 records from one file...
... 50 runs averaged 0.0909136104584 seconds
```