@scottbelden 2017-11-27
I merged https://github.com/tebeka/fastavro/pull/117 which caused a merge conflict here.