@barrywhart 2017-11-27
The merge conflicts have been addressed. This is ready for review.