@barrywhart 2017-11-27
Fixed another merge conflict.