@scottbelden 2017-11-27
I'll try to give this a look over probably tomorrow. At a glance it looks good. I think one initial comment was that perhaps we should move `ReadError` from the common file to the `_reader.pyx` as I think it's only used in that file.