@barrywhart 2017-11-28
Good point about `ReadError` -- I'll change that.

I was thinking, maybe we should update the README with the latest Cython performance numbers. Do you know how the current numbers were generated?