@barrywhart 2017-11-28
I moved the `ReadError` class.