@scottbelden 2017-11-28
I'm not sure how the performance numbers were generated. @tebeka would have more of an idea but I'm assuming whatever script was used might not exist anymore. I did a quick modification of my benchmark script to time it against the standard avro implementation and with the cython code (master, not this PR) I was seeing writes at 14 times faster and reads at 8 times faster. With pure python I see write and reads about 2-3 times faster that the standard avro. I will probably commit this script and the other benchmark script at some point so there is something easily accessible in the future, but until then here is what I have: https://gist.github.com/scottbelden/4556268d0ac5ed39025d29ae0cc715b8