@barrywhart 2017-11-28
Great, I'll try and run that script. If I don't get around to it, it's not a big deal to me; I'm fine with merging without changing the readme.