@barrywhart 2017-11-28
The build failure on Python 3.4 may be an occurrence of this [Cython bug](https://github.com/cython/cython/issues/1894). I'll try the C macro workaround suggested thereto see if it helps.