@barrywhart 2017-11-28
I think I have resolved the Cython issue with Python 3.4. I downgraded to Cython 0.25 as suggested [here](https://github.com/openai/mujoco-py/issues/131#issuecomment-332636429). Hopefully at some point they'll get this working. The specific problem is a bit beyond me unless I take a deep dive into the generated C code.

Given that:
* Python 3.5 was released over 2 years ago
* This Cython issue does not arise in any other Python version (2.7, 3.5, 3.6)

it may be worth considering dropping support for Python 3.4.