@scottbelden 2017-11-28
Is this only happening because of the move of the `ReadError` class? If so maybe it's better to keep it in the common folder so that we can use the more recent version of cython.

As for dropping support for 3.4, I wouldn't have any qualms, but 3.4 is still supported until 2019 (https://docs.python.org/devguide/#status-of-python-branches) so unless it causes major headaches we should probably keep it. I'm not sure how to get statistics on what versions are downloaded, but if we did that and found that 3.4 isn't downloaded much then that could be an argument for dropping support as well.