@barrywhart 2017-11-28
I don't think it has anything to do with moving `ReadError`. I'm seeing the problem with the new `ReaderBase` class, *probably* (I'm guessing) because it is the first "native" (i.e. `cdef`) class in the file.

I did some tinkering to try and work around the problem with C preprocessor settings and/or a couple of "wild guess" tweaks to `_reader.pyx`. None of these helped, but downgrading did.

Here are a few possibilities to consider:
* `fastavro` requires Cython 0.25 at installation time.
  * Pros: Simple, works with all Python versions
  * Cons: Higher potential for Cython version incompatibilities with other libraries
* On Python 3.4, `fastavro` requires Cython 0.25 at installation time. Other Python versions use any version (0.27.3 or later)
  * Pros: Probably reduces risk of Cython version incompatibilities since we're only requiring the older version on Python 3.4
  * Cons: Same as above, but with lower risk. `setup.py` is a little more complex, as it has to check the Python interpreter version.
* Do not distribute the `.pyx` files, only distribute the generated `.c` files
  * Pros: This is the recommended way of distributing Cython extensions. Because Cython does not have to be installed in the target environment, there is no risk of Cython version issues arising (either with the generated C code being bad or with different libraries wanting different Cython versions installed)
  * Cons: Will require tweaking `setup.py` and possibly `publish.sh` to generate the `.c` files and omit the `.pyx` files from the source package. I haven't done this before, so it'd require a bit of tinkering, but probably no more than a couple hours work.