@barrywhart 2017-11-28
BTW, I computed PyPI download stats for `fastavro`. Google makes this information available on BigQuery. There are a few hundred 3.4 downloads in the past month. Pretty light, but there are some.

```
file_project,details_python,downloads
fastavro,2.7.12,12754
fastavro,,12286
fastavro,2.7.11,2918
fastavro,2.7.10,2660
fastavro,3.5.2,1153
fastavro,3.5.4,1131
fastavro,3.6.2,1079
fastavro,2.7.6,932
fastavro,3.6.3,671
fastavro,2.7.13,590
fastavro,3.4.3,381
fastavro,3.6.1,375
fastavro,2.7.5,282
fastavro,2.7,253
fastavro,2.7.14,224
fastavro,2.7.9,214
fastavro,3.5.3,119
fastavro,3.5.1,81
fastavro,3.4.2,77
fastavro,3.6.0,75
fastavro,3.6,48
fastavro,2.7.12+,23
fastavro,3.4.5,20
fastavro,2.6.6,19
fastavro,3.4.0,11
fastavro,3.5,9
fastavro,3.5.0,8
fastavro,3.4.1,6
fastavro,2.7.8,4
fastavro,3.5.4rc1,4
fastavro,3.6.0rc1,3
fastavro,3.4.7,2
fastavro,2.7.0,1
fastavro,3.3.5,1
fastavro,2.6,1
fastavro,2.7.3,1
fastavro,2.7.1,1
fastavro,3.5.2+,1
fastavro,3.4.4,1
```

This is the query I ran:
```
SELECT
  file.project,
  details.python,
  COUNT(*) downloads
FROM
  TABLE_DATE_RANGE(
    [the-psf:pypi.downloads],
    DATE_ADD(CURRENT_TIMESTAMP(), -31, "day"),
    DATE_ADD(CURRENT_TIMESTAMP(), -1, "day")
  )
WHERE
  file.project = 'fastavro'
GROUP BY file.project, details.python
ORDER BY downloads desc
LIMIT 1000
```