@barrywhart 2017-11-29
I think the "hidden" import error bug is fixed. The build is running slowly, but has passed on Python 3.6 and both PyPy versions. Fingers crossed...