@scottbelden 2017-11-29
Thanks again! Merging this and then will try to get a release out shortly.